Refactoring: class GeoDataPointmap::BreakLineAddPointCommand moved to independent files.
